handle config without matching class; simplify tests
