+++ src/test/java/org/zalando/baigan/e2e/s3repo/SomeConfiguration.java
@@ -1,9 +1,10 @@
-package org.zalando.baigan.e2e.simples3repo;
+package org.zalando.baigan.e2e.s3repo;
 
 import org.zalando.baigan.annotation.BaiganConfig;
 
 @BaiganConfig
-public interface SomePlainConfiguration {
+public interface SomeConfiguration {
+    SomeConfigObject someConfig();
     String someValue();
     Boolean isThisTrue();
 }
